profiling: Add Task/GC/Inference/Codegen signposts for Tracy
This adds a variety of metadata to send to Tracy during profiling, in
particular method instance signatures, the Heap size, and a couple of
other small pieces of relevant metadata.

It also adds full support for Task execution, tracking these as "Fibers"
on the Tracy side. Unfortunately, enabling this is blocked on a bug on the
Tracy side. The options for resolution are:
  A. Disable any use of Fibers. Without fibers, asynchronous workloads that
     generate zones across threads are unsupported and will in general break
     the trace.
  B. Disable `TRACY_ON_DEMAND`, which means that profiling data will
     queue up from program start, even if a client is not connected.
  C. Try to implement the missing support for ON_DEMAND + FIBERS upstream.

(A) is what this PR does for now. Thankfully, instrumented calls in the
run-time generally don't span threads so things are at least functional
this way, but it'd be very nice to have a proper fix.