lowering: Don't implicitly `@nospecialize` macros
There's not much purpose of this de-specialization, since macros are
generally excluded from inference / compilation anyway (except for
https://github.com/JuliaLang/julia/pull/57833 which was a bug)

Since our AST is already relatively homogeneous, it's not clear that
we gained much by this de-specialization when it was used. Arbitrarily
de-specializing a function also often leads to bad (wide) edges in
inference etc., causing frequent invalidations.